Fix speculation for if without else clause
If there is no else clause the false-edge leads to a block that will
also be reached through the true-edge namely the continuation block.
Modeling speculation inside this continuation block is wrong. An
analogous issue can occur for the true-edge.
To fix this issue we add proper true- and false-blocks to model
speculation.